PIL-1421: Implemented PR feedback to clean up code
