--- app/controllers/obligationsandsubmissions/ObligationsAndSubmissionsController.scala
@@ -1,56 +1,0 @@
-/*
- * Copyright 2024 HM Revenue & Customs
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package controllers.obligationsandsubmissions
-
-import controllers.actions._
-import controllers.routes.JourneyRecoveryController
-import play.api.Logging
-import play.api.i18n.I18nSupport
-import play.api.mvc.{Action, AnyContent, MessagesControllerComponents}
-import services.obligationsandsubmissions.ObligationsAndSubmissionsService
-import uk.gov.hmrc.play.bootstrap.frontend.controller.FrontendBaseController
-
-import java.time.LocalDate
-import javax.inject.{Inject, Named}
-import scala.concurrent.{ExecutionContext, Future}
-import scala.util.Try
-
-class ObligationsAndSubmissionsController @Inject() (
-  val controllerComponents:               MessagesControllerComponents,
-  getSubscriptionData:                    SubscriptionDataRetrievalAction,
-  requireSubscriptionData:                SubscriptionDataRequiredAction,
-  obligationAndSubmissionsService:        ObligationsAndSubmissionsService,
-  @Named("EnrolmentIdentifier") identify: IdentifierAction
-)(implicit ec:                            ExecutionContext)
-    extends FrontendBaseController
-    with I18nSupport
-    with Logging {
-
-  def onPageLoad(fromDate: String, toDate: String): Action[AnyContent] =
-    (identify andThen getSubscriptionData andThen requireSubscriptionData).async { implicit request =>
-      implicit val pillar2Id: String = request.subscriptionLocalData.plrReference
-      Try {
-        val fromLocalDate = LocalDate.parse(fromDate)
-        val toLocalDate   = LocalDate.parse(toDate)
-
-        obligationAndSubmissionsService
-          .handleData(fromLocalDate, toLocalDate)
-          .map(_ => Redirect(controllers.routes.UnderConstructionController.onPageLoad))
-          .recover { case _: Exception => Redirect(JourneyRecoveryController.onPageLoad()) }
-      }.getOrElse(Future.successful(Redirect(JourneyRecoveryController.onPageLoad())))
-    }
-}
