+++ conf/app.routes
@@ -73,6 +73,4 @@ GET         /below-threshold-notification/confirmation                          
 GET         /below-threshold-notification/problem-with-service                    controllers.btn.BTNProblemWithServiceController.onPageLoad
 GET         /below-threshold-notification/confirmation/download-pdf               controllers.btn.BTNConfirmationController.onDownloadConfirmation
 
-GET         /obligations-and-submissions                                          controllers.obligationsandsubmissions.ObligationsAndSubmissionsController.onPageLoad(fromDate: String, toDate: String)
-
 GET         /submission-history                                                   controllers.submissionhistory.SubmissionHistoryController.onPageLoad
